+++ packages/common/log.js
@@ -6,8 +6,12 @@ const isString = require('lodash.isstring');
 
 function logger() {
   return new Logger({
+    asyncOperationId: process.env.ASYNCOPERATIONID,
     executions: process.env.EXECUTIONS,
+    granules: process.env.GRANULES,
+    parentArn: process.env.PARENTARN,
     sender: process.env.SENDER,
+    stackName: process.env.STACKNAME,
     version: process.env.TASKVERSION
   });
 }